refactor: move `events.spec` properly
